requirements.txt
python-dotenv~=1.0.0
python-telegram-bot==20.4
langchain~=0.0.277
torchaudio~=2.0.2
transformers~=4.32.1
torch~=2.0.1
soundfile~=0.12.1
datasets~=2.14.4